backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
python-multipart>=0.0.6
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0
crewai>=0.1.0
pydantic>=2.0.0
typing-extensions>=4.5.0
# LLM support
langchain-huggingface>=0.0.2  # Updated package for HuggingFacePipeline
langchain-community>=0.0.1
langchain>=0.0.267
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.21.0
bitsandbytes>=0.41.0
# OpenAI support
openai>=1.1.0
# Hashing and validation
python-Levenshtein>=0.21.0
# PDF processing
pypdf>=3.15.1
# For logging and debugging
python-json-logger>=2.0.7